Refactoring Upload Details
