+++ expo/navigation/MainTabNavigator.ts
@@ -6,7 +6,7 @@ import UploadDetails from '../screens/UploadDetails';
 
 const HomeStack = createStackNavigator(
     {
-        Home: UploadDetails,
+        Home: HomeScreen,
         Upload: UploadScreen,
         UploadDetails: UploadDetails,
     },
