+++ expo/screens/UploadDetails.tsx
@@ -1,28 +1,20 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { connect } from 'react-redux';
-import { StyleSheet, View, TouchableOpacity } from 'react-native';
-import { Icon, Input } from 'react-native-elements';
+import { StyleSheet, View } from 'react-native';
 import {
     NavigationScreenProps,
     NavigationScreenOptions,
     NavigationScreenComponent,
     NavigationParams,
 } from 'react-navigation';
-import FullWidthImage from 'react-native-fullwidth-image';
 // @ts-ignore
 import { vw, vh } from 'react-native-expo-viewport-units';
-import { Formik, Field } from 'formik';
 import { ThunkDispatch } from 'redux-thunk';
-import config from '../config';
+import { CapaAutoCompleteProvider } from '../components/CapaAutoCompleteProvider';
 import CapaAutoComplete from '../components/CapaAutoComplete';
-import CapaPhotoSettingsFooter from '../components/CapaPhotoSettingsFooter';
-import {
-    CapaAutoCompleteProvider,
-    useAutoCompleteContext,
-} from '../components/CapaAutoCompleteProvider';
+import PhotoDetailsForm from '../components/CapaPhotoDetailsForm';
 import { S3ActionTypes } from '../modules/s3/types/actions';
 import { Colors, Container, InputField } from '../styles';
-import registerGfx from '../assets/images/bp.jpg';
 
 const styles = StyleSheet.create({
     inputContainerStyle: {
@@ -45,14 +37,6 @@ const styles = StyleSheet.create({
         paddingLeft: 10,
         fontSize: 10,
     },
-    imageView: {
-        flex: 1,
-        width: vw(90),
-        flexDirection: 'column',
-        justifyContent: 'center',
-        marginRight: vw(5),
-        marginLeft: vw(5),
-    },
     avatarContainerStyle: {
         marginTop: 10,
         marginRight: 10,
@@ -73,118 +57,19 @@ interface Item {
     avatar?: string;
 }
 
-interface FormValues {
-    [key: string]: Item;
-}
-
 interface NavStatelessComponent extends React.StatelessComponent {
     navigationOptions?: NavigationScreenOptions;
 }
 
-const renderField = ({
-    field, // { name, value, onChange, onBlur }
-    ...props
-}: {
-    field: object;
-}) => {
-    // @ts-ignore
-    const { value } = props;
-    return (
-        <Input
-            placeholderTextColor="white"
-            inputContainerStyle={styles.inputContainerStyle}
-            containerStyle={[InputField.inputContainer]}
-            inputStyle={[Colors.whiteText, InputField.inputText]}
-            keyboardAppearance="dark"
-            leftIcon={<Icon name="search" color="#fff" />}
-            leftIconContainerStyle={{
-                marginLeft: 0,
-                paddingRight: 5,
-                display: value ? 'none' : 'flex',
-            }}
-            {...field}
-            {...props}
-        />
-    );
-};
-
-export const UploadDetailsForm = () => {
-    const suggestionsContext = useAutoCompleteContext();
-    const [activeTab, setActiveTab] = useState<string | null>(null);
-    const onFocusHandle = (type: string, apiUrl: string) => {
-        suggestionsContext.setActiveUrl(apiUrl);
-        suggestionsContext.setEditMode(true);
-        if (type === 'location') {
-            suggestionsContext.setMapMode(true);
-        } else {
-            suggestionsContext.setMapMode(false);
-        }
-        suggestionsContext.setActive(type);
-    };
-    return !suggestionsContext.editMode ? (
-        <View>
-            <View style={styles.imageView}>
-                {
-                    // @ts-ignore
-                    <FullWidthImage source={registerGfx} ratio={10 / 16} />
-                }
-            </View>
-            <Formik<FormValues | {}>
-                enableReinitialize
-                initialValues={suggestionsContext.form}
-                onSubmit={values => console.log(values)}
-            >
-                {({ values }: { values: FormValues }) => (
-                    <View style={[styles.container]}>
-                        {activeTab === 'Film' && (
-                            <Field
-                                onFocus={() => {
-                                    onFocusHandle('film', `${config.url}/api/film/suggestions`);
-                                }}
-                                value={values.film.name}
-                                label="FILM"
-                                name="Film"
-                                component={renderField}
-                                placeholder=""
-                            />
-                        )}
-                        {activeTab === 'Camera' && (
-                            <Field
-                                onFocus={() => {
-                                    onFocusHandle('gear', `${config.url}/api/camera/suggestions`);
-                                }}
-                                value={values.gear.name}
-                                label="CAMERA"
-                                name="Gear"
-                                component={renderField}
-                                placeholder=""
-                            />
-                        )}
-                        {activeTab === 'Location' && (
-                            <Field
-                                onFocus={() => {
-                                    onFocusHandle('location', `${config.url}/api/location`);
-                                }}
-                                value={values.location.name}
-                                label="LOCATION"
-                                name="Location"
-                                component={renderField}
-                                placeholder=""
-                            />
-                        )}
-                    </View>
-                )}
-            </Formik>
-            <CapaPhotoSettingsFooter changeActiveTab={tab => setActiveTab(tab)} />
-        </View>
-    ) : null;
-};
-
-export const UploadDetails: NavigationScreenComponent = () => {
+export const UploadDetails: NavigationScreenComponent = ({
+    navigation,
+    dispatchSaveDetails,
+}: NavigationParams) => {
+    const { photo } = navigation.state.params;
     return (
         <View style={styles.containerNoPadding}>
             <CapaAutoCompleteProvider>
-                <UploadDetailsForm />
+                <PhotoDetailsForm photo={photo} handleSave={dispatchSaveDetails} />
                 <CapaAutoComplete />
             </CapaAutoCompleteProvider>
         </View>
@@ -198,31 +83,11 @@ UploadDetails.navigationOptions = ({ navigation }: NavigationParams): Navigation
         marginRight: 15,
         borderBottomWidth: 0,
     },
-    headerLeft: (
-        <Icon
-            name="close"
-            color="#fff"
-            onPress={(): void => {
-                navigation.goBack();
-            }}
-            Component={TouchableOpacity}
-        />
-    ),
-    headerRight: (
-        <Icon
-            name="check"
-            color="#fff"
-            onPress={(): void => {
-                navigation.goBack();
-            }}
-            Component={TouchableOpacity}
-        />
-    ),
 });
 
 const mapDispatchToProps = (dispatch: ThunkDispatch<{}, {}, S3ActionTypes>): DispatchProps => ({
-    dispatchSaveDetails: () => {
-        console.log(dispatch);
+    dispatchSaveDetails: data => {
+        console.log(data);
     },
 });
